Move all mock files into internal/mocks and use mock prefix
